feat: add vector storage and semantic workspace context search
Enable pgvector-backed semantic search, context priorities, embedding generation, and local daemon/MCP tooling so agents can retrieve ranked workspace context.

+++ src/Espada.Domain/SeedWork/IHasConcurrencyVersion.cs
@@ -2,5 +2,5 @@ namespace Espada.Domain.SeedWork;
 
 public interface IHasConcurrencyVersion
 {
-    long Version { get; }
+    uint Version { get; }
 }
